% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ACEsearch.connect}
\alias{ACEsearch.connect}
\title{Connects to ACE instance}
\usage{
ACEsearch.connect(url)
}
\arguments{
\item{url}{url address of a running ACE instance, usually containing port information}
}
\value{
data frame containing connection information used for all other accessory functions
}
\description{
Attempts to connect to ACE instance using URL:PORT
}
\examples{
\donttest{ACEsearch.connect("http://localhost:8080")}

}
