% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ACEsearch.contains}
\alias{ACEsearch.contains}
\title{Returns whether patient with the specified patient_id exists}
\usage{
ACEsearch.contains(connection, patient_id)
}
\arguments{
\item{connection}{connection object returned from connect(url) function}

\item{patient_id}{numerical id of the patient}
}
\value{
TRUE or FALSE
}
\description{
Returns the statistics information
}
\examples{
\donttest{ACEsearch.contains(ACEsearch.connect('http://localhost:8080'), 123)}


}
