% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ACEsearch.query}
\alias{ACEsearch.query}
\title{Queries ACE and returns a list of patient IDs}
\usage{
ACEsearch.query(connection, query, output_time = FALSE)
}
\arguments{
\item{connection}{connection object returned from connect(url) function}

\item{query}{ACE query}

\item{output_time}{equivalent to wrapping the query in OUTPUT() command. Togerther with patient IDs outputs each
time interval in patient's timeline when the query was evaluated as true}
}
\value{
data frame containing patient IDs and time intervals (optional)
}
\description{
Queries ACE and returns a list of patient IDs
}
\examples{
\donttest{ACEsearch.query(ACEsearch.connect('http://localhost:8080'), 'ICD9=250.50')}
\donttest{ACEsearch.query(ACEsearch.connect('http://localhost:8080'), 'ICD9=250.50', TRUE)}


}
