% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ACEsearch.status}
\alias{ACEsearch.status}
\title{Returns the status of the ACE search engine}
\usage{
ACEsearch.status(connection)
}
\arguments{
\item{connection}{connection object returned from connect(url) function}
}
\value{
data frame containing patient IDs and time intervals (optional)
}
\description{
Returns the status of the ACE search engine
}
\examples{
\donttest{ACEsearch.status(ACEsearch.connect('http://localhost:8080'))}


}
